fungi-classifier-app/requirements.txt
flask==2.0.1
pillow==9.0.0
transformers==4.15.0
requests==2.26.0
python-dotenv==0.19.2
werkzeug==2.0.2
